moving models from the `domain` module to the `features:clients:domain` module
